requirement.txt
Flask
Flask_Session
psycopg2
py-bcrypt
qrcode
pillow
waitress